// src_no/types/nav.types.ts
export type PageId =
  | 'dashboard'
  | 'master'
  | 'entry'
  | 'masters-config'
  | 'report'
  | 'opinion-poll'

export type EntryModuleId =
  | 'voter'
  | 'booth'
  | 'volunteer'
  | 'event'
  | 'campaign'
  | 'user'
  | 'warroom'
  | 'dashboard'
  | 'alliance'
  | 'keypeople'
  | 'feedback'
  | 'commitment'
  | 'grievance'
  | 'volunteer-activity'
  | 'agent-activity'
  | 'field-activity'
  | 'voter-survey'

export type MasterModuleId =
  | 'area'
  | 'booth'
  | 'village'
  | 'scheme'
  | 'issue'
  | 'candidate'
  | 'party'